Extract x86-specific `TaskContext`
