+++ framework/aster-frame/src/task/mod.rs
@@ -12,5 +12,5 @@ pub use self::{
     priority::Priority,
     processor::{current_task, disable_preempt, preempt, schedule, DisablePreemptGuard},
     scheduler::{add_task, set_scheduler, FifoScheduler, Scheduler},
-    task::{Task, TaskAdapter, TaskOptions, TaskStatus},
+    task::{Task, TaskAdapter, TaskContextApi, TaskOptions, TaskStatus},
 };
